Move CircularBuffer to `utils` folder.
