+++ include/jetstream/memory/base.hh
@@ -4,7 +4,6 @@
 #include "jetstream/types.hh"
 #include "jetstream/memory/macros.hh"
 #include "jetstream/memory/types.hh"
-#include "jetstream/memory/buffer.hh"
 #include "jetstream/memory/token.hh"
 
 #ifdef JETSTREAM_BACKEND_CPU_AVAILABLE
