+++ src/memory/meson.build
@@ -3,6 +3,4 @@ subdir('cuda')
 subdir('metal')
 subdir('vulkan')
 
-src_lst += files([
-    'buffer.cc',
-])
+subdir('utils')
