+++ src/memory/utils/meson.build
@@ -1,0 +1,3 @@
+src_lst += files([
+    'circular_buffer.cc',
+])
